requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.7
django-extensions==2.2.5
djangorestframework==3.9.4
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
drfdocs==0.0.11
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Markdown==3.1.1
# pkg-resources==0.0.0
psycopg2-binary==2.8.3
pydot==1.4.1
pyparsing==2.4.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
boto3==1.9.252
botocore==1.12.252
django-storages==1.8.0
jmespath==0.9.4
python-dateutil==2.8.0
s3transfer==0.2.1